+++ app.py
@@ -1,13 +1,3 @@
 from flask import Flask
 
 app = Flask(__name__)
-
-
-@app.route('/')
-def hello():
-    return "Hello World"
-
-
-@app.route('/name/<name>')
-def hello_name(name):
-    return f"Hello {name}"
